requirements.txt
Flask==2.0.1
WTForms~=2.3.3
Flask-WTF==0.15.1
Flask-Login==0.5.0
superannotate==4.1.9
objectpath==0.6.1
passlib==1.7.4
python-dotenv==0.20.0